Redfish API Bindings
The majority of this work was copied from gorackhd. Things were changed in the makefile as well asa the testing to accompany the new spec file. Started with a completely new git log to not interfere with gorackhd moving forward. Big thanks to @akutz for his work on getting it to this point where it was easily portable

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>